+++ Samples~/DBSamples/Runtime/Item DB/InventoryEditorPreset.cs
@@ -1,4 +1,4 @@
-#if DEBUG
+#if UNITY_EDITOR
 using System;
 using System.Collections;
 using System.Collections.Generic;